+++ endpoint/endpoint_test.go
@@ -23,4 +23,12 @@ func TestNewEndpoint(t *testing.T) {
 	if e.DNSName != "example.org" || e.Target != "1.2.3.4" {
 		t.Error("endpoint is not initialized correctly")
 	}
+	if e.Labels == nil {
+		t.Error("Labels is not initialized")
+	}
+
+	w := NewEndpoint("example.org.", "load-balancer.com.")
+	if w.DNSName != "example.org" || w.Target != "load-balancer.com" {
+		t.Error("endpoint is not initialized correctly")
+	}
 }