+++ pkg/apis/externaldns/types.go
@@ -25,27 +25,28 @@ import (
 )
 
 var (
-	defaultHealthPort = "9090"
-	defaultLogFormat  = "text"
+	defaultMetricsAddress = ":7979"
+	defaultLogFormat      = "text"
 )
 
 // Config is a project-wide configuration
 type Config struct {
-	InCluster     bool
-	KubeConfig    string
-	Namespace     string
-	Zone          string
-	Sources       []string
-	Provider      string
-	GoogleProject string
-	Policy        string
-	HealthPort    string
-	Interval      time.Duration
-	Once          bool
-	DryRun        bool
-	Debug         bool
-	LogFormat     string
-	Version       bool
+	InCluster      bool
+	KubeConfig     string
+	Namespace      string
+	Zone           string
+	Sources        []string
+	Provider       string
+	GoogleProject  string
+	Policy         string
+	Compatibility  bool
+	MetricsAddress string
+	Interval       time.Duration
+	Once           bool
+	DryRun         bool
+	Debug          bool
+	LogFormat      string
+	Version        bool
 }
 
 // NewConfig returns new Config object
@@ -64,7 +65,8 @@ func (cfg *Config) ParseFlags(args []string) error {
 	flags.StringVar(&cfg.Provider, "provider", "", "the DNS provider to materialize the records in")
 	flags.StringVar(&cfg.GoogleProject, "google-project", "", "gcloud project to target")
 	flags.StringVar(&cfg.Policy, "policy", "sync", "the policy to use. options: [\"sync\", \"upsert-only\"]")
-	flags.StringVar(&cfg.HealthPort, "health-port", defaultHealthPort, "health port to listen on")
+	flags.BoolVar(&cfg.Compatibility, "compatibility", false, "enable to process annotation semantics from legacy implementations")
+	flags.StringVar(&cfg.MetricsAddress, "metrics-address", defaultMetricsAddress, "address to expose metrics on")
 	flags.StringVar(&cfg.LogFormat, "log-format", defaultLogFormat, "log format output. options: [\"text\", \"json\"]")
 	flags.DurationVar(&cfg.Interval, "interval", time.Minute, "interval between synchronizations")
 	flags.BoolVar(&cfg.Once, "once", false, "run once and exit")