+++ source/service_test.go
@@ -40,6 +40,7 @@ func testServiceEndpoints(t *testing.T) {
 		targetNamespace string
 		svcNamespace    string
 		svcName         string
+		compatibility   bool
 		annotations     map[string]string
 		lbs             []string
 		expected        []*endpoint.Endpoint
@@ -49,6 +50,7 @@ func testServiceEndpoints(t *testing.T) {
 			"",
 			"testing",
 			"foo",
+			false,
 			map[string]string{},
 			[]string{"1.2.3.4"},
 			[]*endpoint.Endpoint{},
@@ -58,8 +60,9 @@ func testServiceEndpoints(t *testing.T) {
 			"",
 			"testing",
 			"foo",
+			false,
 			map[string]string{
-				hostnameAnnotationKey: "foo.example.org",
+				hostnameAnnotationKey: "foo.example.org.",
 			},
 			[]string{"1.2.3.4"},
 			[]*endpoint.Endpoint{
@@ -71,11 +74,27 @@ func testServiceEndpoints(t *testing.T) {
 			"",
 			"testing",
 			"foo",
+			false,
 			map[string]string{
-				hostnameAnnotationKey: "foo.example.org",
+				hostnameAnnotationKey: "foo.example.org.",
 			},
-			[]string{"lb.example.com"},
+			[]string{"lb.example.com"}, // Kubernetes omits the trailing dot
 			[]*endpoint.Endpoint{
+				{DNSName: "foo.example.org", Target: "lb.example.com"},
+			},
+		},
+		{
+			"annotated services can omit trailing dot",
+			"",
+			"testing",
+			"foo",
+			false,
+			map[string]string{
+				hostnameAnnotationKey: "foo.example.org", // Trailing dot is omitted
+			},
+			[]string{"1.2.3.4", "lb.example.com"}, // Kubernetes omits the trailing dot
+			[]*endpoint.Endpoint{
+				{DNSName: "foo.example.org", Target: "1.2.3.4"},
 				{DNSName: "foo.example.org", Target: "lb.example.com"},
 			},
 		},
@@ -84,9 +103,10 @@ func testServiceEndpoints(t *testing.T) {
 			"",
 			"testing",
 			"foo",
+			false,
 			map[string]string{
 				controllerAnnotationKey: controllerAnnotationValue,
-				hostnameAnnotationKey:   "foo.example.org",
+				hostnameAnnotationKey:   "foo.example.org.",
 			},
 			[]string{"1.2.3.4"},
 			[]*endpoint.Endpoint{
@@ -98,9 +118,10 @@ func testServiceEndpoints(t *testing.T) {
 			"",
 			"testing",
 			"foo",
+			false,
 			map[string]string{
 				controllerAnnotationKey: "some-other-tool",
-				hostnameAnnotationKey:   "foo.example.org",
+				hostnameAnnotationKey:   "foo.example.org.",
 			},
 			[]string{"1.2.3.4"},
 			[]*endpoint.Endpoint{},
@@ -110,8 +131,9 @@ func testServiceEndpoints(t *testing.T) {
 			"testing",
 			"testing",
 			"foo",
+			false,
 			map[string]string{
-				hostnameAnnotationKey: "foo.example.org",
+				hostnameAnnotationKey: "foo.example.org.",
 			},
 			[]string{"1.2.3.4"},
 			[]*endpoint.Endpoint{
@@ -123,8 +145,9 @@ func testServiceEndpoints(t *testing.T) {
 			"testing",
 			"other-testing",
 			"foo",
+			false,
 			map[string]string{
-				hostnameAnnotationKey: "foo.example.org",
+				hostnameAnnotationKey: "foo.example.org.",
 			},
 			[]string{"1.2.3.4"},
 			[]*endpoint.Endpoint{},
@@ -134,8 +157,9 @@ func testServiceEndpoints(t *testing.T) {
 			"",
 			"other-testing",
 			"foo",
+			false,
 			map[string]string{
-				hostnameAnnotationKey: "foo.example.org",
+				hostnameAnnotationKey: "foo.example.org.",
 			},
 			[]string{"1.2.3.4"},
 			[]*endpoint.Endpoint{
@@ -147,8 +171,9 @@ func testServiceEndpoints(t *testing.T) {
 			"",
 			"testing",
 			"foo",
+			false,
 			map[string]string{
-				hostnameAnnotationKey: "foo.example.org",
+				hostnameAnnotationKey: "foo.example.org.",
 			},
 			[]string{},
 			[]*endpoint.Endpoint{},
@@ -158,13 +183,40 @@ func testServiceEndpoints(t *testing.T) {
 			"",
 			"testing",
 			"foo",
+			false,
 			map[string]string{
-				hostnameAnnotationKey: "foo.example.org",
+				hostnameAnnotationKey: "foo.example.org.",
 			},
 			[]string{"1.2.3.4", "8.8.8.8"},
 			[]*endpoint.Endpoint{
 				{DNSName: "foo.example.org", Target: "1.2.3.4"},
 				{DNSName: "foo.example.org", Target: "8.8.8.8"},
+			},
+		},
+		{
+			"services annotated with legacy mate annotations are ignored in default mode",
+			"",
+			"testing",
+			"foo",
+			false,
+			map[string]string{
+				"zalando.org/dnsname": "foo.example.org.",
+			},
+			[]string{"1.2.3.4"},
+			[]*endpoint.Endpoint{},
+		},
+		{
+			"services annotated with legacy mate annotations return an endpoint in compatibility mode",
+			"",
+			"testing",
+			"foo",
+			true,
+			map[string]string{
+				"zalando.org/dnsname": "foo.example.org.",
+			},
+			[]string{"1.2.3.4"},
+			[]*endpoint.Endpoint{
+				{DNSName: "foo.example.org", Target: "1.2.3.4"},
 			},
 		},
 	} {
@@ -201,7 +253,7 @@ func testServiceEndpoints(t *testing.T) {
 			}
 
 			// Create our object under test and get the endpoints.
-			client := NewServiceSource(kubernetes, tc.targetNamespace)
+			client := NewServiceSource(kubernetes, tc.targetNamespace, tc.compatibility)
 
 			endpoints, err := client.Endpoints()
 			if err != nil {
@@ -222,7 +274,7 @@ func BenchmarkServiceEndpoints(b *testing.B) {
 			Namespace: "testing",
 			Name:      "foo",
 			Annotations: map[string]string{
-				hostnameAnnotationKey: "foo.example.org",
+				hostnameAnnotationKey: "foo.example.org.",
 			},
 		},
 		Status: v1.ServiceStatus{
@@ -240,7 +292,7 @@ func BenchmarkServiceEndpoints(b *testing.B) {
 		b.Fatal(err)
 	}
 
-	client := NewServiceSource(kubernetes, v1.NamespaceAll)
+	client := NewServiceSource(kubernetes, v1.NamespaceAll, false)
 
 	for i := 0; i < b.N; i++ {
 		_, err := client.Endpoints()